requirements.txt
biopython>=1.71
bcbio-gff>=0.6.4
pandas>=0.23.0
numpy>=1.14.3